+++ Engine/Plugins/PluginsManager.cs
@@ -222,6 +222,7 @@ namespace JocysCom.VS.AiCompanion.Engine
 					ai.Item = item;
 					// Map Text, Audio and Video converter methods.
 					mm.VideoToText = ai.VideoToText;
+					mm.GetStructuredImageAnalysisInstructions = () => Global.AppSettings.StructuredImageAnalysisInstructions;
 					mm.AISpeakCallback = Global.AvatarOptionsPanel.AI_SpeakSSML;
 					//mm.CaptureCameraImageCallback = CameraHelper.CaptureCameraImage;
 					methodResult = await InvokeMethod(methodInfo, mm, invokeParams, cancellationTokenSource.Token);